rename folders to snake_case
