+++ tests/books_page/test_books_page.py
@@ -2,13 +2,11 @@ __author__ = 'Alexandra Dunga'
 
 import pytest
 
-from Config.config import TestData
-from Tests.test_base import BaseTest
-from Pages.books_page import BooksPage
+from config.config import TestData
+from tests.test_base import BaseTest
+from pages.books_page import BooksPage
 
 
-# @pytest.mark.usefixtures("setup_books_page")
-# How to use a class fixture without passing it as a parameter in the test methods?
 class TestBooks(BaseTest):
 
     """Check if the signup link exists"""
