+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
-from Config.config import TestData
-from Pages.books_page import BooksPage
-from Pages.login_page import LoginPage
+from config.config import TestData
+from pages.books_page import BooksPage
+from pages.login_page import LoginPage
 
 
 @pytest.fixture(scope="class")
