+++ tests/home_page/test_home_page.py
@@ -1,7 +1,7 @@
 __author__ = 'Alexandra Dunga'
 
-from Tests.test_base import BaseTest
-from Config.config import TestData
+from tests.test_base import BaseTest
+from config.config import TestData
 
 
 class TestHome(BaseTest):
