+++ tests/login_page/test_login_page.py
@@ -1,7 +1,7 @@
 __author__ = 'Alexandra Dunga'
 
-from Config.config import TestData
-from Tests.test_base import BaseTest
+from config.config import TestData
+from tests.test_base import BaseTest
 
 
 class TestLogin(BaseTest):
